Revert "Revert "Always get transformed manifest server-side""
