+++ content/webapp/components/IIIFViewer/IIIFViewerImage.tsx
@@ -1,7 +1,7 @@
 import { forwardRef, useState } from 'react';
 import styled from 'styled-components';
 import { convertIiifImageUri } from '@weco/common/utils/convert-image-uri';
-import { convertRequestUriToInfoUri } from '@weco/content/utils/convert-iiif-uri';
+import { convertRequestUriToInfoUri } from '@weco/content/utils/iiif/convert-iiif-uri';
 async function getImageMax(url: string): Promise<number> {
   try {
     const infoUrl = convertRequestUriToInfoUri(url);
