+++ content/webapp/test/utils/convert-iiif-uri.test.ts
@@ -1,4 +1,4 @@
-import { convertRequestUriToInfoUri } from '../../utils/convert-iiif-uri';
+import { convertRequestUriToInfoUri } from '@weco/content/utils/iiif/convert-iiif-uri';
 
 describe('convertRequestUriToInfoUri', () => {
   it('finds the info.json for a IIIF URI', () => {
