+++ content/webapp/utils/iiif/convert-iiif-uri.ts
