+++ toggles/README.md
@@ -5,6 +5,7 @@ different cohorts of people and stakeholders safely and incrementally.
 
 There is [a great article by Martin Fowler][martin-fowler-feature-toggles] on the subject.
 
+## Categories
 We currently use three categories of toggles:
 
 ### 1. Feature development toggle
